+++ @internal/ssr/src/classes/fetch/request/FetchRequest.js
@@ -1,0 +1,8 @@
+class FetchRequest {
+  constructor(url, config) {
+    this.url = url;
+    this.config = config;
+  }
+}
+
+export { FetchRequest };
