+++ @internal/ssr/src/classes/fetch/response/Response.js
@@ -1,0 +1,15 @@
+import { SyncPromise } from '../../base/SyncPromise';
+
+class Response {
+  constructor(ctx) {
+    this.json = () => new SyncPromise(resolve => {
+      resolve(ctx.getData());
+    });
+
+    this.text = () => new SyncPromise(resolve => {
+      resolve(ctx.getData());
+    });
+  }
+}
+
+export { Response };
