+++ @internal/ssr/src/classes/fetch/response/__tests__/Response.spec.js
@@ -1,0 +1,44 @@
+import { Response } from '../Response';
+
+describe('Response', () => {
+  it('should be a class', () => {
+    expect(Response).toBeAClass();
+  });
+
+  describe('instance', () => {
+    let mockContext;
+
+    let instance;
+    beforeEach(() => {
+      mockContext = {
+        getData: jest.fn(),
+      };
+
+      instance = new Response(mockContext);
+    });
+
+    describe('json', () => {
+      it('should be a function', () => {
+        expect(instance.json).toBeInstanceOf(Function);
+      });
+
+      it('should return the data in the value slot wrapped in a promise', () => {
+        const mockData = {};
+        mockContext.getData.mockReturnValue(mockData);
+        instance.json().then(data => expect(data).toBe(mockData));
+      });
+    });
+
+    describe('text', () => {
+      it('should be a function', () => {
+        expect(instance.text).toBeInstanceOf(Function);
+      });
+
+      it('should return the data in the value slot wrapped in a promise', () => {
+        const mockData = 'data';
+        mockContext.getData.mockReturnValue(mockData);
+        instance.text().then(data => expect(data).toBe(mockData));
+      });
+    });
+  });
+});
